python/src/multi_agent_orchestrator/retrievers/retriever.py
from abc import ABC, abstractmethod

class Retriever(ABC):
    """
    Abstract base class for Retriever implementations.
    This class provides a common structure for different types of retrievers.
    """

    def __init__(self, options: dict):
        """
        Constructor for the Retriever class.

        Args:
            options (dict): Configuration options for the retriever.
        """
        self._options = options

    @abstractmethod
    async def retrieve(self, text: str) -> any:
        """
        Abstract method for retrieving information based on input text.
        This method must be implemented by all concrete subclasses.

        Args:
            text (str): The input text to base the retrieval on.

        Returns:
            any: The retrieved information.
        """
        pass

    @abstractmethod
    async def retrieve_and_combine_results(self, text: str) -> any:
        """
        Abstract method for retrieving information and combining results.
        This method must be implemented by all concrete subclasses.
        It's expected to perform retrieval and then combine or process the results in some way.

        Args:
            text (str): The input text to base the retrieval on.

        Returns:
            any: The combined retrieval results.
        """
        pass

    @abstractmethod
    async def retrieve_and_generate(self, text: str) -> any:
        """
        Abstract method for retrieving information and generating something based on the results.
        This method must be implemented by all concrete subclasses.
        It's expected to perform retrieval and then use the results to generate new information.

        Args:
            text (str): The input text to base the retrieval on.

        Returns:
            any: The generated information based on retrieval results.
        """
        pass