python/src/tests/retrievers/test_retriever.py
import pytest
from asyncio import Future

# Assume the Retriever class is in a file named retriever.py
from multi_agent_orchestrator.retrievers import Retriever

class ConcreteRetriever(Retriever):
    """A concrete implementation of Retriever for testing purposes."""

    async def retrieve(self, text: str) -> str:
        return f"Retrieved: {text}"

    async def retrieve_and_combine_results(self, text: str) -> str:
        return f"Combined: {text}"

    async def retrieve_and_generate(self, text: str) -> str:
        return f"Generated: {text}"

@pytest.fixture
def retriever():
    """Fixture to create a ConcreteRetriever instance."""
    return ConcreteRetriever({"option1": "value1"})

@pytest.mark.asyncio
async def test_retrieve(retriever):
    result = await retriever.retrieve("test")
    assert result == "Retrieved: test"

@pytest.mark.asyncio
async def test_retrieve_and_combine_results(retriever):
    result = await retriever.retrieve_and_combine_results("test")
    assert result == "Combined: test"

@pytest.mark.asyncio
async def test_retrieve_and_generate(retriever):
    result = await retriever.retrieve_and_generate("test")
    assert result == "Generated: test"

def test_init():
    options = {"option1": "value1", "option2": "value2"}
    retriever = ConcreteRetriever(options)
    assert retriever._options == options

def test_abstract_class():
    with pytest.raises(TypeError):
        Retriever({})

@pytest.mark.asyncio
async def test_abstract_methods():
    class IncompleteRetriever(Retriever):
        async def retrieve(self, text: str) -> str:
            return ""

    with pytest.raises(TypeError):
        IncompleteRetriever({})